feat: add PSR-4 autoloading
