--- includes/index.php
@@ -1,2 +1,0 @@
-<?php
-//I have known the silence of the stars and of the sea
